+++ eslint.config.js
@@ -18,7 +18,7 @@ export default tseslint.config(
     },
   },
   {
-    files: ['scripts/**/*.mjs', 'eslint.config.js'],
+    files: ['**/scripts/**/*.mjs', 'eslint.config.js'],
     languageOptions: {
       globals: {
         console: 'readonly',